# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Book:
    properties:
      author:
        type: string
      created_at:
        description: '[NEW] Waktu dibuat'
        type: string
      id:
        type: integer
      stock:
        description: '[NEW] Menyimpan jumlah stok'
        type: integer
      title:
        type: string
      updated_at:
        description: '[NEW] Waktu terakhir diedit'
        type: string
    type: object
...
  models.Loan:
    properties:
      book:
        allOf:
        - $ref: '#/definitions/models.Book'
        description: Relasi ke Book
      book_id:
        description: ID Buku
        type: integer
      created_at:
        type: string
      id:
        type: integer
      loan_date:
        description: Tanggal Pinjam
        type: string
      return_date:
        description: Tanggal Harus Kembali
        type: string
      status:
        description: 'Status: borrowed/returned'
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relasi ke User
      user_id:
        description: ID Peminjam
        type: integer
    type: object
  models.LoanRequest:
    properties:
      book_id:
        type: integer
    required:
    - book_id
    type: object
...
  models.User:
    properties:
      address:
        type: string
      born_date:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Tim Developer
    url: http://www.example.com/support
  description: Ini adalah dokumentasi API server untuk manajemen User, Buku, dan Peminjaman.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gin Gorm API (Manajemen Buku)
  version: "2.0"
paths:
  /admin/loans:
    get:
      consumes:
      - application/json
      description: Menampilkan seluruh data peminjaman dari semua user. Khusus Admin.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Loan'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Bukan Admin)
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lihat Semua Peminjaman (Admin)
      tags:
      - admin
...
  /loans:
    post:
      consumes:
      - application/json
      description: Membuat data peminjaman baru dan mengurangi stok buku.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Peminjaman (Book ID)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Loan'
        "400":
          description: Stok habis atau sedang meminjam
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Buku tidak ditemukan
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Pinjam Buku
      tags:
      - loans
  /loans/my:
    get:
      consumes:
      - application/json
      description: Menampilkan daftar semua buku yang pernah dipinjam oleh user yang
        sedang login.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Loan'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lihat Riwayat Peminjaman Saya
      tags:
      - loans
  /loans/return/{id}:
    post:
      consumes:
      - application/json
      description: Mengubah status peminjaman menjadi 'returned' dan menambah stok
        buku. Hanya pemilik peminjaman yang boleh akses.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Loan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Pesan Sukses
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Sudah dikembalikan
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Bukan milik user ini
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Data Loan tidak ditemukan
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Kembalikan Buku
      tags:
      - loans
...
securityDefinitions:
  BearerAuth:
    description: Masukkan token dengan format "Bearer <token_anda>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"